create RefreshChat component / Do not disable when it’s a proximity chat room
